director page ui
