+++ src/Components/Director_verified_applications.js
@@ -1,7 +1,6 @@
 import React, { Component, useEffect, useState, useContext } from "react";
-import { BrowserRouter as Router, useNavigate } from "react-router-dom";
+import { BrowserRouter as Router, useNavigate, Link } from "react-router-dom";
 
-import { Button, Form, FormGroup } from "react-bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { AuthContext } from "./Auth";
 import "./Home_authority.css";
@@ -9,6 +8,36 @@ import { Container, Row, Col, Alert, Breadcrumb, Card } from "react-bootstrap";
 
 function Director_verified_applications() {
   const email = localStorage.getItem("email");
+  const [searchTerm, setSearchTerm] = useState("");
+  const [sortColumn, setSortColumn] = useState("id");
+  const [sortDirection, setSortDirection] = useState("asc");
+  const [data, setData] = useState([]);
+
+  const handleSearchChange = (event) => {
+    setSearchTerm(event.target.value);
+  };
+
+  const handleSortChange = (event) => {
+    setSortColumn(event.target.value);
+    setSortDirection(sortDirection === "asc" ? "desc" : "asc");
+  };
+
+  const filteredData = data.filter((row) =>
+    Object.values(row).some((value) =>
+      value.toString().toLowerCase().includes(searchTerm.toLowerCase())
+    )
+  );
+
+  const sortedData = filteredData.sort((a, b) => {
+    const isAsc = sortDirection === "asc";
+    if (a[sortColumn] < b[sortColumn]) {
+      return isAsc ? -1 : 1;
+    } else if (a[sortColumn] > b[sortColumn]) {
+      return isAsc ? 1 : -1;
+    } else {
+      return 0;
+    }
+  });
 
   let user_data = {
     email: email,
@@ -25,10 +54,23 @@ function Director_verified_applications() {
       }
     );
 
-    const data = await res.json();
-    console.log(data["result"]);
+    const data2 = await res.json();
+    console.log(data2["result"]);
 
-     setresult_arr(data["result"]);
+    setresult_arr(data2["result"]);
+
+    const updateData = [];
+    data2["result"].map((id1) => {
+      console.log(id1[0]);
+      updateData.push({
+        id: parseInt(id1[0]),
+        amount: parseInt(JSON.parse(id1[1]).user.netAmntClaimed),
+        date: JSON.parse(id1[1]).user.date,
+        status: id1[2],
+      });
+      console.log(data.length);
+    });
+    setData(updateData.reverse());
   };
   useEffect(() => {
     getApplicationId();
@@ -48,14 +90,9 @@ function Director_verified_applications() {
   };
 
   return (
-    <div>
-      <div id="top_navbar">
-        <div id="name">Welcome</div>
-        <div id="email">{email}</div>
-      </div>
-
+    <div style={{ display: "flex" }}>
       <div
-        id="sidebar"
+        id="sidebar1"
         class="d-flex flex-column  flex-shrink-0 p-3 text-white"
       >
         <a href="#" class="text-white text-decoration-none">
@@ -92,35 +129,60 @@ function Director_verified_applications() {
           </li>
         </ul>
       </div>
-      <div id="last_heading">
-        <h4>Verfied applications </h4>
-        <h6>(applications which are approved by you will appear here)</h6>
-      </div>
-      <div className="application_list">
-        {result_arr.map((id) => (
-          <div
-            className="application_id"
-            onClick={() => {
-              navigate("ShowAllApplication/" + id);
-            }}
-            style={{ cursor: "pointer" }}
-          >
-            <Container>
-              <Row>
-                <h6>Application Id : {id[0]}</h6>
-                {console.log()}
-                <div
-                  style={{ display: "flex", justifyContent: "space-between" }}
-                >
-                  {/* <div>Date of application : {}</div>
-                  <div>
-                    Amount claimed : {id[1]}
-                  </div> */}
-                </div>
-              </Row>
-            </Container>
+      <div className="main-content">
+        <div className="top-navbar">
+          <div className="welcome">
+            <div className="welcome-icon">
+              <i className="fas fa-user-circle"></i>{" "}
+              {/* Add margin to move the icon */}
+            </div>
+
+            <div className="welcome-text">
+              <div className="name">Director</div>{" "}
+              {/* Replace [Dummy Name] with "Mohit" */}
+              <div className="email">
+                <i className="fas fa-envelope"></i> {email}{" "}
+                {/* You can use envelope icon for email */}
+              </div>
+            </div>
           </div>
-        ))}
+        </div>
+        <hr></hr>
+        <div id="last_heading">
+          <h4>Verfied applications </h4>
+          <h6>(applications which are approved by you will appear here)</h6>
+        </div>
+        <div className="application-list">
+          <table className="table">
+            <thead>
+              <tr>
+                <th>Application ID</th>
+                <th>Amount Claimed</th>
+                <th>Date of submission</th>
+                <th>Action</th> {/* Added Action column */}
+              </tr>
+            </thead>
+            <tbody>
+              {data.map((row, index) => (
+                <tr key={index}>
+                  <td>Application {row.id}</td>
+                  <td>{row.amount}</td>
+                  <td>{row.date}</td>
+                  <td>
+                    <button
+                      onClick={() => {
+                        navigate("ShowAllApplication/" + row.id);
+                      }}
+                      className="btn btn-primary"
+                    >
+                      View Application
+                    </button>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
       </div>
     </div>
   );
